Move symbols for internal methods to abstract-ops
