+++ src/lib/abstract-ops/internal-methods.ts
@@ -1,2 +1,4 @@
+export const AbortSteps = Symbol('[[AbortSteps]]');
+export const ErrorSteps = Symbol('[[ErrorSteps]]');
 export const CancelSteps = Symbol('[[CancelSteps]]');
 export const PullSteps = Symbol('[[PullSteps]]');
